pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comatch"
version = "0.1.0"
description = "Official implementation of CoMatch: Dynamic Covisibility-Aware Transformer for bilateral subpixel-level semi-dense image matching"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12,<3.13"
dependencies = [
	"albumentations>=1.4,<2",
	"h5py>=3.10,<4",
	"joblib>=1.3,<2",
	"kornia>=0.7,<0.8",
	"loguru>=0.7,<1",
	"matplotlib>=3.8,<4",
	"numpy>=1.26,<3",
	"opencv-python>=4.10,<5",
	"pillow>=10,<12",
	"pytorch-lightning>=2.2,<3",
	"tensorboard>=2.16,<3",
	"torch>=2.2,<3",
	"torchmetrics>=1.3,<2",
	"tqdm>=4.66,<5",
	"yacs>=0.1.8,<0.2",
]

[dependency-groups]
dev = [
	"autopep8>=2.3,<3",
	"ipython>=8.22,<9",
	"jupyterlab>=4.1,<5",
	"pylint>=3,<4",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["comatch*", "src*", "configs*"]

[tool.uv]
package = true
